feat: share place chat link with place members
