+++ src/dotnet/Chat.UI.Blazor/Components/Share/share-modal.css
@@ -11,8 +11,11 @@
     @apply overflow-y-hidden;
 }
 .share-modal .form-block {
-    @apply p-0;
+    @apply py-0;
     @apply overflow-y-hidden;
+}
+.share-modal.share-modal-with-contact-selector .form-block {
+    @apply px-0;
 }
 .share-modal .no-padding-header-decorator {
     @apply mx-0;
