+++ src/dotnet/Chat.UI.Blazor/Services/EditMembersUI.cs
@@ -9,7 +9,7 @@ public class EditMembersUI(ChatUIHub hub)
         if (chat.IsPublicPlaceChat())
             return false;
 
-        var provider = new ChatMemberSelector(Hub, chat.Id);
+        var provider = new NewChatMemberSelector(Hub, chat.Id);
         var selected = await provider.ListPreSelectedUserIds(default).ConfigureAwait(false);
         var available = await provider.ListUserIds(default).ConfigureAwait(false);
         var canAddContacts = available.Except(selected).Any();
