+++ src/dotnet/UI.Blazor/Components/Share/ShareModalModel.cs
@@ -4,4 +4,9 @@ public sealed record ShareModalModel(
     ShareKind Kind,
     string Title,
     string TargetTitle,
-    ShareRequest Request);
+    ShareRequest Request,
+    IShareModalSelector? SelectorPrefs);
+
+public interface IShareModalSelector;
+
+public record PrivatePlaceMembersShareSelector(ChatId ChatId) : IShareModalSelector;
