+++ src/dotnet/UI.Blazor/Services/ShareUI.cs
@@ -7,5 +7,5 @@ public sealed class ShareUI(UIHub hub) : ScopedServiceBase<UIHub>(hub)
     public Task<ModalRef> Share(ShareModalModel model)
         => ModalUI.Show(model);
     public Task<ModalRef> Share(ShareKind kind, string title, string targetTitle, ShareRequest request)
-        => ModalUI.Show(new ShareModalModel(kind, title, targetTitle, request));
+        => ModalUI.Show(new ShareModalModel(kind, title, targetTitle, request, null));
 }
